(PC-17104)[API] feat: remove database subscription_message
persisted subscription_message are irrelevant because we cannot trust them: the latest is used as it may not be the right message to show to the user; a deprecated message could remain and may not be applicable anymore
