+++ api/src/pcapi/admin/install.py
@@ -4,7 +4,6 @@ from flask import Flask
 from flask_admin.base import Admin
 from sqlalchemy.orm.session import Session
 
-from pcapi.admin.custom_views import inapp_messages
 from pcapi.admin.custom_views import offer_view
 from pcapi.admin.custom_views.admin_user_view import AdminUserView
 from pcapi.admin.custom_views.allocine_pivot_view import AllocinePivotView
@@ -42,7 +41,6 @@ from pcapi.core.payments.models import CustomReimbursementRule
 from pcapi.core.providers.models import AllocinePivot
 from pcapi.core.providers.models import Provider
 from pcapi.core.providers.models import VenueProvider
-from pcapi.core.subscription import models as subscription_models
 from pcapi.core.users.models import User
 from pcapi.core.users.models import UserEmailHistory
 from pcapi.models.beneficiary_import import BeneficiaryImport
@@ -278,14 +276,6 @@ def install_views(admin: Admin, session: Session) -> None:
         SubcategoryView(
             name="Sous-catégories",
             endpoint="subcategories",
-            category=Category.CUSTOM_OPERATIONS,
-        )
-    )
-    admin.add_view(
-        inapp_messages.MessageView(
-            subscription_models.SubscriptionMessage,
-            session,
-            name="Messages utilisateurs dans l'app",
             category=Category.CUSTOM_OPERATIONS,
         )
     )
